Fix: Null id in location header
