+++ app/actions/Read.java
@@ -384,19 +384,4 @@ public class Read extends RegalAction {
 	return linkedObjects;
     }
 
-    private String createAggregationUri(String pid) {
-	return Globals.useHttpUris ? "http://" + Globals.server + "/resource/"
-		+ pid : pid;
-    }
-
-    /**
-     * @param node
-     *            the node
-     * @return the http address of the resource
-     */
-    public String getHttpUriOfResource(Node node) {
-	return Globals.useHttpUris ? node.getAggregationUri() : "http://"
-		+ Globals.server + "/resource/" + node.getAggregationUri();
-    }
-
 }
